+++ client/index.html
@@ -118,11 +118,8 @@
   <script type="text/javascript" src="dashboard/dashboardController.js"></script>
   <script type="text/javascript" src="signin/signinController.js"></script>
   <script type="text/javascript" src="leaderboard/leaderBoardController.js"></script>
-
   <script type="text/javascript" src="services/leaderBoardService.js"></script>
   <script type="text/javascript" src="services/dashboardService.js"></script>
-
-
   <script type="text/javascript" src="news/newsController.js"></script>
