+++ server/config/middleware.js
@@ -10,8 +10,7 @@ var Transaction = require('../../db/models').Transaction;
 var Room_user = require('../../db/models').Room_user;
 var expressJwt = require('express-jwt');
 var jwt = require('jsonwebtoken');
-var Twit = require('twit');
-var moment = require('moment')
+
 
 
 module.exports = function (app, express) {
@@ -21,6 +20,7 @@ module.exports = function (app, express) {
   var leagueRouter = express.Router();
   var portfolioRouter = express.Router();
   var stockRouter = express.Router();
+  var tweetRouter = express.Router();
 
   // Configuring middleware
   app.use(bodyParser.urlencoded({ extend: true }));
@@ -87,32 +87,7 @@ module.exports = function (app, express) {
   // })
 
 
-var T= new Twit({
-    consumer_key: 'wQX96fSNXd5Oyo0mhw0AHto5u'
-  , consumer_secret: 'shbjbyIGPVpopdw5FD3o1EtQNUwbAe5aVkkKg7pziLHPlnBH9x'
-  , app_only_auth: true
-});
 
-var params={ q: '$F', count: 5 }
-app.get('/api/tweets', function(req,res){
-  console.log('here');
-  
-  
-    var messages = [];
-    
-  T.get('search/tweets', params, function(err,data,response){
-    for(var i=0;i<data.statuses.length;i++){
-        var status=data.statuses[i];
-        var tweetDate = status.created_at
-        var date = new Date(Date.parse(tweetDate.replace(/( \+)/, ' UTC$1')));
-        var time = moment(date).fromNow();
-        messages.push({text:status.text,
-                  user: status.user.screen_name,
-                  created_at: time})
-    }
-    res.json(messages)
-  })
-})
 
 
 
@@ -126,6 +101,9 @@ app.get('/api/tweets', function(req,res){
   app.use('/api/leagues', leagueRouter);
   app.use('/api/portfolios', portfolioRouter);
   app.use('/api/stocks', stockRouter);
+  app.use('/api/tweets', tweetRouter);
+
+  require('../tweets/tweetRoutes.js')(tweetRouter);
 
   require('../stocks/stockRoutes.js')(stockRouter);
