+++ server/server.js
@@ -9,5 +9,4 @@ app.listen(port, function () {
   console.log('Connected to port: ' + port);
 });
 
-
 module.exports = app;
