+++ server/tweets/tweetRoutes.js
@@ -1,0 +1,7 @@
+var tweetController = require('./tweetController.js');
+
+module.exports = function(app){
+  
+  app.get('/', tweetController.getTweets)
+
+}
